+++ src/App.jsx
@@ -1,15 +1,16 @@
 import {Composition} from 'remotion';
 import {MyVideo} from './components/video';
+import {fps, screenSize} from './constants';
 export const App = () => {
 	return (
 		<>
 			<Composition
 				id="Video"
 				component={MyVideo}
-				durationInFrames={4000}
-				fps={60}
-				width={1920}
-				height={1080}
+				durationInFrames={2000}
+				fps={fps}
+				width={screenSize[0]}
+				height={screenSize[1]}
 			/>
 		</>
 	);