+++ CHANGELOG.md
@@ -14,3 +14,12 @@
 - Implementation audio in background
 - Mute all videos
 - Implementation simple animations & style to videos
+
+
+###### July 19, 2022
+- Add Animations And Transitions
+- Create New Components Like [Bar,Card,Gradient,Label]
+- Create Animation For End Scene With Use [Bar] Component
+- Dynamicly change color - music - fps - screen size and etc width [config.json]
+- System Testing Remotation Package
+- Finished Project